Implemented access tokens.
It is now possible to generate access tokens. An access token allows a user to temporarily act on behalf of another user. An access token grants a user the same permissions to a webstrate as the user who generated the access token. Access tokens only grant access to the webstrate they were issued for.

Access tokens are immediately expired/revoked when the permissions of a webstrate changes or a document gets restored (as this may also change permissions). It is not otherwise possible to expire access tokens (other than by reaching the end of their duration). Making it possible to expire access tokens manually will be implemented on request.

An access token is generated by performing a POST request to the webstrate in question. The POST request must contain a `token` field. The value of the `token` field will be used as the token duration in seconds. If the value is not a positive integer, it defaults to 300 seconds.

A token request response will look similar to:

    {
      "webstrateId": "test-token",
      "username": "kbadk",
      "provider": "github",
      "token": "Hyg57Qg7Z",
      "expiration": 1497540063
    }

Access tokens cannot be generated from access token-based access (i.e. you cannot generate tokens using tokens).

To access a webstrate using an access token, put the token in the query string. For the above example:

    /test-token/?token=Hyg57Qg7Z

A list of all access tokens can be viewed by accessing /<webstrateId>/?tokens.