Improvements to websocket reuse. Including the same document twice now works, as well as transcluding a document that doesn't exist.
Having multiple subscriptions to the same webstrate causes ShareDB to behave oddly and cut off parts of operations for (so far) unknown reasons, so we now force the creation of a new websocket when transcluding the same webstrate multiple times. This will rarely be the case, so for the vast majority of applications, the websocket will still be reused.