Fixed attribute diff'ing issue. When ops were coming in, we didn't execute them immediately (i.e. in the same tick), but rather we just executed them as a new task (using setImmediate). This was supposed to be an optimization.
However, when a bunch of ops come in at the same time, and there are MutationObservers at play in the DOM, the MutationObservers may trigger actions inbetween the received operations are applied to the DOM, because the MutationObservers create microtasks. If these microtasks then modify an attribute that's in the midst of being applied operations to, it won't necessarily be possible to apply the remaining received operations.

The solution is to execute the received operations immediately after each other in the current task (i.e. not in new tasks). Fortunately, this is possible, because these operations were created (by Google's DiffMatchPatch) right after each other, and thus sent as a collection of operations. I do fear that a similar issue could turn up in the future when MutationObservers modify too rapidly after having received ops.