Implemented `webstrate.dataSaved()` which returns a promise that gets resolved when all inflight ops have been saved on the server. If an application wants to make changes to the DOM before navigating away, it previously had to more or less guess then DOM changes had actually been saved on the server, either by doing hacky things such as sending a signal and waiting for the reply (which wasn't even guaranteed to work), or simply adding a delay.
We can now listen for the acknowledgement from the sever that the ops have been applied with `webstrate.dataSaved()`. This function returns a promise that gets resolved when the client has confirmation that it's up to date with the server, i.e. that the server has received the client's ops (and there are no other ops in queue) and saved them.

		const text = document.createTextNode("Hello world");
		box.appendChild(text);
		await webstrate.dataSaved();
		console.log("Our data has now been registered and it\'s safe to close the page!");