Implements protected mode.
Protected mode is enabled by adding data-protected as an attribute on the body tag and reloading(!) the page. Afterwards, all attributes set and elements inserted will automatically be transient. To create an element that's approved to be in the DOM (i.e. not transient), the element should be created with:

	document.createElement(tagName, { approved: true });

Likewise, for setting a non-transient attribute:

	Element.setAttribute(name, value, { approved: true });