Made it possible to override `isTransientElement` and `isTransientAttribute` from within a webstrate (+ improved transient tests, including tests for this). `isTransientElement` and `isTransientAttribute` are now exposed on `webstrate.config`. To make all attributes beginning with `custom-` transient, as well as make all elements with tag name `custom` transient, one could do:
    webstrate.on('loaded', () => {
        const isTransientAttribute = webstrate.config.isTransientAttribute;
        webstrate.config.isTransientAttribute = (DOMNode, attributeName) =>
            attributeName.startsWith('custom-') || isTransientAttribute(DOMNode, attributeName);

        const isTransientElement = webstrate.config.isTransientElement;
        webstrate.config.isTransientElement = (DOMNode) =>
            DOMNode.tagName.toLowerCase() === 'custom' || isTransientElement(DOMNode);
    });

Inclusion of the existing `isTransientElement` and `isTransientAttribute` functions in the re-implementations is not required, but recommended, as to not conflict with the existing transient attributes and elements.

Redefining the properties on only select clients of a webstrate may result in undefined behaviour. Redefining the properties after having already modified the DOM is also undefined behaviour. Basically, anything but redefining the propeties immediately after page load on all clients is considered undefined behaviour.